Refactor compilation caching
