+++ tests/helpers/test_cache.py
@@ -2,6 +2,7 @@ import os
 import tempfile
 
 from sinol_make.helpers import compile
+from sinol_make.helpers.cache import check_compiled, save_compiled
 
 
 def test_compilation_caching():
@@ -10,22 +11,22 @@ def test_compilation_caching():
         program = os.path.join(tmpdir, 'program.cpp')
         open(program, 'w').write('int main() { return 0; }')
 
-        assert compile.check_compiled(program) is None
+        assert check_compiled(program) is None
 
         assert compile.compile(program, os.path.join(tmpdir, 'program'), compile_log=None)
-        exe_path = compile.check_compiled(program)
+        exe_path = check_compiled(program)
         assert exe_path is not None
 
         assert compile.compile(program, os.path.join(tmpdir, 'program'), compile_log=None)
-        exe_path2 = compile.check_compiled(program)
+        exe_path2 = check_compiled(program)
         assert exe_path2 == exe_path
 
         open(program, 'w').write('int main() { return 1; }')
-        assert compile.check_compiled(program) is None
+        assert check_compiled(program) is None
         assert compile.compile(program, os.path.join(tmpdir, 'program'), compile_log=None)
-        assert compile.check_compiled(program) is not None
+        assert check_compiled(program) is not None
 
         open(program, 'w').write('int main() { return 0; }')
-        assert compile.check_compiled(program) is None
+        assert check_compiled(program) is None
         assert compile.compile(program, os.path.join(tmpdir, 'program'), compile_log=None)
-        assert compile.check_compiled(program) is not None
+        assert check_compiled(program) is not None
