+++ tests/helpers/test_compile.py
@@ -1,7 +1,7 @@
 import os
 import tempfile
 
-from sinol_make.helpers import compile
+from sinol_make.helpers.cache import save_compiled, check_compiled
 
 
 def test_compilation_caching():
@@ -12,7 +12,7 @@ def test_compilation_caching():
         with open(os.path.join(os.getcwd(), "test.e"), "w") as f:
             f.write("")
 
-        assert compile.check_compiled(os.path.join(os.getcwd(), "test.txt")) is None
-        compile.save_compiled(os.path.join(os.getcwd(), "test.txt"),
-                              os.path.join(os.getcwd(), "test.e"))
-        assert compile.check_compiled(os.path.join(os.getcwd(), "test.txt")) == os.path.join(os.getcwd(), "test.e")
+        assert check_compiled(os.path.join(os.getcwd(), "test.txt")) is None
+        save_compiled(os.path.join(os.getcwd(), "test.txt"),
+                                               os.path.join(os.getcwd(), "test.e"))
+        assert check_compiled(os.path.join(os.getcwd(), "test.txt")) == os.path.join(os.getcwd(), "test.e")
